최종 프로젝트 업로드

+++ src/main/java/com/neighbor21/ggits/web/service/systemmng/UserMngService.java
@@ -7,13 +7,20 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.neighbor21.ggits.common.entity.MOpAuthority;
 import com.neighbor21.ggits.common.entity.MOpCode;
 import com.neighbor21.ggits.common.entity.MOpGrpInfo;
+import com.neighbor21.ggits.common.entity.MOpLayoutMstInfo;
+import com.neighbor21.ggits.common.entity.MOpMenu;
 import com.neighbor21.ggits.common.entity.MOpOperator;
+import com.neighbor21.ggits.common.entity.MOpUserCntnSystemMenu;
+import com.neighbor21.ggits.common.enums.CntnSystemCd;
+import com.neighbor21.ggits.common.enums.LayoutMenuInfo;
 import com.neighbor21.ggits.common.enums.OprtrGrd;
 import com.neighbor21.ggits.common.mapper.LOpPgmLognMapper;
 import com.neighbor21.ggits.common.mapper.LOpUseMenuMapper;
@@ -21,9 +28,13 @@ import com.neighbor21.ggits.common.mapper.LTcSrvrLogInfoMapper;
 import com.neighbor21.ggits.common.mapper.MOpAuthorityMapper;
 import com.neighbor21.ggits.common.mapper.MOpCodeMapper;
 import com.neighbor21.ggits.common.mapper.MOpGrpInfoMapper;
+import com.neighbor21.ggits.common.mapper.MOpLayoutMstInfoMapper;
+import com.neighbor21.ggits.common.mapper.MOpMenuMapper;
 import com.neighbor21.ggits.common.mapper.MOpOperatorMapper;
 import com.neighbor21.ggits.common.mapper.MOpOperatorMenuMapper;
+import com.neighbor21.ggits.common.mapper.MOpUserCntnSystemMenuMapper;
 import com.neighbor21.ggits.common.util.GgitsCommonUtils;
+import com.neighbor21.ggits.common.util.LoginSessionUtils;
 import com.neighbor21.ggits.common.util.PasswordUtils;
 import com.neighbor21.ggits.support.exception.CommonException;
 import com.neighbor21.ggits.support.exception.ErrorCode;
@@ -31,6 +42,8 @@ import com.neighbor21.ggits.support.exception.ErrorCode;
 @Service
 public class UserMngService{
 	
+	@Autowired
+	MOpMenuMapper mOpMenuMapper;
 
 	@Autowired
 	MOpOperatorMapper mOpOperatorMapper;
@@ -56,6 +69,12 @@ public class UserMngService{
 	@Autowired
 	MOpCodeMapper mopCodeMapper;
 	
+	@Autowired
+	MOpLayoutMstInfoMapper mOpLayoutMstInfoMapper;
+	
+	@Autowired
+	MOpUserCntnSystemMenuMapper mOpUserCntnSystemMenuMapper;
+	
 	/**
 	  * @Method Name : findOneUserDetailByOprtrId
 	  * @작성일 : 2023. 8. 28.
@@ -80,6 +99,14 @@ public class UserMngService{
 				userDetailInfo.setCdNm(dbMOpCode.getCdNm());				
 			}
 		}
+		if(!GgitsCommonUtils.isNull(userDetailInfo.getMngInstCd())) {
+			MOpCode mOpCode = new MOpCode();
+			mOpCode.setCdId(userDetailInfo.getMngInstCd());
+			MOpCode dbMOpCode = mopCodeMapper.findOneMOpCodeByCdId(mOpCode);
+			if(mOpCode != null) {
+				userDetailInfo.setMngInstNm(dbMOpCode.getCdNm());				
+			}
+		}
 		
 		return userDetailInfo;
 	}
@@ -93,21 +120,34 @@ public class UserMngService{
 	  */
 	public void updateUserInfo(MOpOperator mOpOperator) {
 		
+		//권한 체크 로직
+		MOpOperator sessionMOpOperator = LoginSessionUtils.getMOpOperatorInfo();
+		
+		MOpGrpInfo schMOpGrpInfo = new MOpGrpInfo();
+		schMOpGrpInfo.setGrpId(sessionMOpOperator.getGrpId());
+		
+		MOpGrpInfo mOpGrpInfo =  mOpGrpInfoMapper.findOneGroupDetailByGrpId(schMOpGrpInfo);
+		
 		MOpOperator userDetailInfo = mOpOperatorMapper.findOneUserDetailByOprtrId(mOpOperator);
-		if(userDetailInfo == null) {
-			throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"유저 정보가 존재하지 않습니다.");
-		}
-		try {
+		
+		MOpAuthority mOpAuthority = mOpAuthorityMapper.findOneByAuthId(mOpGrpInfo.getAuthId());
+		
+		if("AUC000".equals(mOpAuthority.getAuthCd()) || "SUPER".equals(sessionMOpOperator.getOprtrGrd()) || mOpOperator.getOprtrId() == sessionMOpOperator.getOprtrId()) {
+			if(userDetailInfo == null) {
+				throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"유저 정보가 존재하지 않습니다.");
+			}
+			
 			userDetailInfo.setOprtrNm(mOpOperator.getOprtrNm());
 			userDetailInfo.setAddngCd(mOpOperator.getAddngCd());
 			userDetailInfo.setOprtrTel(mOpOperator.getOprtrTel());
 			userDetailInfo.setGrpId(mOpOperator.getGrpId());
 			userDetailInfo.setOprtrSttsCd(mOpOperator.getOprtrSttsCd());
+			userDetailInfo.setMngInstCd(mOpOperator.getMngInstCd());
 			userDetailInfo.setOprtrTel(GgitsCommonUtils.removeHyphen(mOpOperator.getOprtrTel()));
 			
 			mOpOperatorMapper.update(userDetailInfo);
-		}catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"유저 정보가 수정되지 않았습니다. 관리자에게 문의하세요.");
+		} else {
+			throw new CommonException(ErrorCode.NO_PERMISSION);
 		}
 	}
 	
@@ -119,20 +159,29 @@ public class UserMngService{
 	 * @param mOpOperator
 	 */
 	public void deleteUserInfo(MOpOperator mOpOperator) {
+		//권한 체크 로직
+		MOpOperator sessionMOpOperator = LoginSessionUtils.getMOpOperatorInfo();
 		
-		try {
+		if(sessionMOpOperator != null && "SUPER".equals(sessionMOpOperator.getOprtrGrd())) {
+			if(mOpOperator.getOprtrId() == null) {
+				throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"유저 정보가 삭제되지 않았습니다. 관리자에게 문의하세요.");
+			}
 			
 			long oprtrId = mOpOperator.getOprtrId();
-
+			
 			lOpPgmLognMapper.deleteByOprtrId(oprtrId);
 			
 			lTcSrvrLogInfoMapper.deleteByOprtrId(oprtrId);
-
+			
 			lOpUseMenuMapper.deleteByOprtrId(oprtrId);
 			
+			mOpLayoutMstInfoMapper.deleteByOprtrId(oprtrId);
+
+			mOpUserCntnSystemMenuMapper.deleteByOprtrId(oprtrId);
+
 			mOpOperatorMapper.deleteByOprtrId(oprtrId);
-		}catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"유저 정보가 삭제되지 않았습니다. 관리자에게 문의하세요.");
+		} else {
+			throw new CommonException(ErrorCode.NO_PERMISSION);
 		}
 	}
 
@@ -189,25 +238,60 @@ public class UserMngService{
 		MOpGrpInfo operatorGrpInfo = mOpGrpInfoMapper.findOneMOpGrpInfoForDefaultGrp(mOpGrpInfo);
 		
 		if(GgitsCommonUtils.isNull(operatorGrpInfo.getGrpId())) {
-			throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"일반 관리자 기본 그룹이 존재하지 않습니다. 생성 후 삭제해 주세요.");
+			throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"일반 관리자 기본 그룹이 존재하지 않습니다.");
 		}
 		
-		try {
 			
-			//비밀번호 암호화
-			String oprtrPswd = PasswordUtils.hashPassword(mOpOperator.getOprtrPswd()); 
-			
-			// 유저 상태 코드 
-			String oprtrSttsCd = "OSC001"; //OSC001 : 미승인 , OSC002 : 승인 , OSC003 : 중지 
-			mOpOperator.setOprtrGrd(OprtrGrd.GENERAL_ADMIN.getGrade());
-			mOpOperator.setOprtrPswd(oprtrPswd);
-			mOpOperator.setOprtrSttsCd(oprtrSttsCd);
-			mOpOperator.setGrpId(operatorGrpInfo.getGrpId());
-			mOpOperator.setOprtrTel(GgitsCommonUtils.removeHyphen(mOpOperator.getOprtrTel()));
-			mOpOperatorMapper.save(mOpOperator);
-			
-		} catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_SAVE_FAIL,"관리자 등록 실패했습니다.");    
+		//비밀번호 암호화
+		String oprtrPswd = PasswordUtils.hashPassword(mOpOperator.getOprtrPswd()); 
+		
+		// 유저 상태 코드 
+		String oprtrSttsCd = "OSC001"; //OSC001 : 미승인 , OSC002 : 승인 , OSC003 : 중지 
+		long oprtrId = mOpOperatorMapper.findOneByOprtrIdNextVal();
+		
+		mOpOperator.setOprtrId(oprtrId);
+		mOpOperator.setOprtrGrd(OprtrGrd.GENERAL_ADMIN.getGrade());
+		mOpOperator.setOprtrPswd(oprtrPswd);
+		mOpOperator.setOprtrSttsCd(oprtrSttsCd);
+		mOpOperator.setGrpId(operatorGrpInfo.getGrpId());
+		mOpOperator.setLayoutNo(1L);
+		mOpOperator.setOprtrTel(GgitsCommonUtils.removeHyphen(mOpOperator.getOprtrTel()));
+		mOpOperatorMapper.save(mOpOperator);
+		
+		//레이아웃 디폴트값 추가
+		List<LayoutMenuInfo> layOutMenuList = Stream.of(LayoutMenuInfo.values()).collect(Collectors.toList());
+
+		for(LayoutMenuInfo layoutMenuInfo : layOutMenuList) {
+		   MOpMenu mOpMenu = new MOpMenu();
+		   mOpMenu.setMenuPttrnType(layoutMenuInfo.getCode());
+		   String menuId = mOpMenuMapper.findOneMenuIdByMenuPttrnType(mOpMenu);
+
+		   MOpLayoutMstInfo mOpLayoutMstInfo = new MOpLayoutMstInfo();
+		   if(!GgitsCommonUtils.isNull(menuId)) {
+			   mOpLayoutMstInfo.setMenuId(menuId);
+		   } else {
+			   mOpLayoutMstInfo.setMenuId(layoutMenuInfo.getCode());
+		   }
+		   mOpLayoutMstInfo.setLayoutId(GgitsCommonUtils.getUuid(10));
+		   mOpLayoutMstInfo.setOprtrId(oprtrId);
+		   mOpLayoutMstInfo.setLayoutMenuNm(layoutMenuInfo.getName());
+		   mOpLayoutMstInfo.setLayoutSttsCd("1");
+		   mOpLayoutMstInfo.setLayout1UseYn("Y");
+		   mOpLayoutMstInfo.setLayout2UseYn("N");
+		   mOpLayoutMstInfo.setLayout3UseYn("N");
+		   mOpLayoutMstInfo.setFnctType(layoutMenuInfo.getFncType());
+		   mOpLayoutMstInfo.setDataTypeCd(layoutMenuInfo.getDataTypeCd());
+		   
+		   mOpLayoutMstInfoMapper.saveMOpLayoutMstInfo(mOpLayoutMstInfo);
+		}
+		
+		//대문 메뉴 default값 추가
+		for(CntnSystemCd r : CntnSystemCd.values()) {
+			MOpUserCntnSystemMenu mOpUserCntnSystemMenu = new MOpUserCntnSystemMenu();
+			mOpUserCntnSystemMenu.setOprtrId(oprtrId);
+			mOpUserCntnSystemMenu.setUseYn(r.getDefaultYn());
+			mOpUserCntnSystemMenu.setCntnSystemCd(r.getCode());
+			mOpUserCntnSystemMenuMapper.saveMOpUserCntnSystemMenu(mOpUserCntnSystemMenu);
 		}
 	}
 	
@@ -262,16 +346,11 @@ public class UserMngService{
 			throw new CommonException(ErrorCode.USER_INFO_IS_EXIST);
 		}
 		
-		try {
+		//비밀번호 암호화
+		String oprtrPswd = PasswordUtils.hashPassword(mOpOperator.getOprtrPswd()); 
+		updateUserInfo.setOprtrPswd(oprtrPswd);
+		mOpOperatorMapper.update(updateUserInfo);
 			
-			//비밀번호 암호화
-			String oprtrPswd = PasswordUtils.hashPassword(mOpOperator.getOprtrPswd()); 
-			updateUserInfo.setOprtrPswd(oprtrPswd);
-			mOpOperatorMapper.update(updateUserInfo);
-			
-		} catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_SAVE_FAIL,"관리자 등록 실패했습니다.");    
-		}
 	}
 	/**
 	  * @Method Name : updateGrpInfo
@@ -282,21 +361,17 @@ public class UserMngService{
 	  * @param oprtrIdArr
 	  */
 	public void updateGrpInfo(MOpGrpInfo mOpGrpInfo,String strOprtrIdArr) {
-		try {
-			
-			MOpGrpInfo groupDetail = mOpGrpInfoMapper.findOneGroupDetailByGrpId(mOpGrpInfo);
-			if(groupDetail == null) {
-				throw new CommonException(ErrorCode.ENTITY_DATA_NULL, "그룹 정보가 존재하지 않습니다.");
-			}
-			groupDetail.setGrpNm(mOpGrpInfo.getGrpNm());
-			groupDetail.setGrpDescr(mOpGrpInfo.getGrpDescr());
-			groupDetail.setAuthId(mOpGrpInfo.getAuthId());
-			mOpGrpInfoMapper.update(groupDetail);
-			updateUserGrpInfo(mOpGrpInfo.getGrpId(),strOprtrIdArr,"UPDATE");		 
-			
-		} catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"그룹정보 수정 실패했습니다.");
+		
+		MOpGrpInfo groupDetail = mOpGrpInfoMapper.findOneGroupDetailByGrpId(mOpGrpInfo);
+		
+		if(groupDetail == null) {
+			throw new CommonException(ErrorCode.ENTITY_DATA_NULL, "그룹 정보가 존재하지 않습니다.");
 		}
+		groupDetail.setGrpNm(mOpGrpInfo.getGrpNm());
+		groupDetail.setGrpDescr(mOpGrpInfo.getGrpDescr());
+		groupDetail.setAuthId(mOpGrpInfo.getAuthId());
+		mOpGrpInfoMapper.update(groupDetail);
+		updateUserGrpInfo(mOpGrpInfo.getGrpId(),strOprtrIdArr,"UPDATE");		 
 	}
 	/**
 	 * @Method Name : saveGrpInfo
@@ -320,19 +395,15 @@ public class UserMngService{
 				}
 			}
 		}
-		
-		try {
-			if(GgitsCommonUtils.isNull(mOpGrpInfo.getBscGrpYn())) {
-				mOpGrpInfo.setBscGrpYn("N");
-			}
-			if(GgitsCommonUtils.isNull(mOpGrpInfo.getUpperOprtrAuthGrpYn())) {
-				mOpGrpInfo.setUpperOprtrAuthGrpYn("N");
-			}
-			mOpGrpInfoMapper.save(mOpGrpInfo);
-			updateUserGrpInfo(mOpGrpInfo.getGrpId(),strOprtrIdArr,"SAVE");
-		} catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"그룹정보 등록 실패했습니다.");
+	
+		if(GgitsCommonUtils.isNull(mOpGrpInfo.getBscGrpYn())) {
+			mOpGrpInfo.setBscGrpYn("N");
 		}
+		if(GgitsCommonUtils.isNull(mOpGrpInfo.getUpperOprtrAuthGrpYn())) {
+			mOpGrpInfo.setUpperOprtrAuthGrpYn("N");
+		}
+		mOpGrpInfoMapper.save(mOpGrpInfo);
+		updateUserGrpInfo(mOpGrpInfo.getGrpId(),strOprtrIdArr,"SAVE");
 	}
 	
 	/**
@@ -354,12 +425,8 @@ public class UserMngService{
 			throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"일반 관리자 기본 그룹이 존재하지 않습니다. 생성 후 삭제해 주세요.");
 		}
 		
-		try {
-			updateUserGrpInfo(defaultGeneralGrpInfo.getGrpId(),strOprtrIdArr,"DELETE");
-			mOpGrpInfoMapper.delete(mOpGrpInfo);
-		} catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"그룹정보 삭제 실패했습니다.");
-		}
+		updateUserGrpInfo(defaultGeneralGrpInfo.getGrpId(),strOprtrIdArr,"DELETE");
+		mOpGrpInfoMapper.delete(mOpGrpInfo);
 	}
 	
 	
@@ -372,63 +439,58 @@ public class UserMngService{
 	 */
 	public void updateUserGrpInfo(String grpId,String strOprtrIdArr,String type) {
 		
-		try {
-			String[] oprtrIdArr = null;
-			
-			//UPDATE 시 기존에 DB유저와 파라미터 유저 비교 후 없으면 기본 일반 그룹으로 GRP_ID 수정
-			List<MOpOperator> userList = new ArrayList<MOpOperator>();
-			List<Long> deleteOprtrIdList = new ArrayList<Long>();
-			if("UPDATE".equals(type)) {
-				MOpOperator mOpOperator = new MOpOperator();
-				mOpOperator.setGrpId(grpId);
-				userList = mOpOperatorMapper.findAllUserListByGrpId(mOpOperator);
-				//추가할 사용자가 존재하지 않을경우 
+		String[] oprtrIdArr = null;
+		
+		//UPDATE 시 기존에 DB유저와 파라미터 유저 비교 후 없으면 기본 일반 그룹으로 GRP_ID 수정
+		List<MOpOperator> userList = new ArrayList<MOpOperator>();
+		List<Long> deleteOprtrIdList = new ArrayList<Long>();
+		if("UPDATE".equals(type)) {
+			MOpOperator mOpOperator = new MOpOperator();
+			mOpOperator.setGrpId(grpId);
+			userList = mOpOperatorMapper.findAllUserListByGrpId(mOpOperator);
+			//추가할 사용자가 존재하지 않을경우 
+			if(userList.size() > 0) {
+				deleteOprtrIdList = userList.stream().map(x -> x.getOprtrId()).collect(Collectors.toList());  
+			}
+		}
+		
+		//선택된 사용자 ID 목록 
+		if(!strOprtrIdArr.isEmpty()) {
+			oprtrIdArr = strOprtrIdArr.split(",");
+			if(oprtrIdArr.length > 0) {
+				
+				//그룹에 추가 할 유저 목록 OpOperator Update
+				List<Long> updateOprtrIdList = Arrays.stream(oprtrIdArr).filter(x -> !GgitsCommonUtils.isNull(x)).map(Long::parseLong).collect(Collectors.toList());
+				Map<String,Object> grpInfoUpdateMap = new HashMap<String, Object>();
+				
+				grpInfoUpdateMap.put("grpId", grpId);
+				grpInfoUpdateMap.put("updateOprtrIdList", updateOprtrIdList);
+				
+				mOpOperatorMapper.updateForGrpInfo(grpInfoUpdateMap);
+				
+				//UPDATE 시 기존에 DB유저와 파라미터 유저 비교 후 없으면 기본 일반 그룹으로 GRP_ID 수정
 				if(userList.size() > 0) {
-					deleteOprtrIdList = userList.stream().map(x -> x.getOprtrId()).collect(Collectors.toList());  
+					deleteOprtrIdList = userList.stream().filter(x -> !updateOprtrIdList.contains(x.getOprtrId())).map(x -> x.getOprtrId()).collect(Collectors.toList());
 				}
 			}
+		}
+		
+		//삭제할 목록이 존재하면 MOpOperator Update
+		if(deleteOprtrIdList.size() > 0) {
+			//기본 그룹, 일반 관리자 그룹 DB 조회
+			MOpGrpInfo defaultGrpInfo = new MOpGrpInfo();
+			defaultGrpInfo.setBscGrpYn("Y");//기본 그룹
+			defaultGrpInfo.setUpperOprtrAuthGrpYn("N");//일반 관리자 그룹
 			
-			//선택된 사용자 ID 목록 
-			if(!strOprtrIdArr.isEmpty()) {
-				oprtrIdArr = strOprtrIdArr.split(",");
-				if(oprtrIdArr.length > 0) {
-					
-					//그룹에 추가 할 유저 목록 OpOperator Update
-					List<Long> updateOprtrIdList = Arrays.stream(oprtrIdArr).filter(x -> !GgitsCommonUtils.isNull(x)).map(Long::parseLong).collect(Collectors.toList());
-					Map<String,Object> grpInfoUpdateMap = new HashMap<String, Object>();
-					
-					grpInfoUpdateMap.put("grpId", grpId);
-					grpInfoUpdateMap.put("updateOprtrIdList", updateOprtrIdList);
-					
-					mOpOperatorMapper.updateForGrpInfo(grpInfoUpdateMap);
-					
-					//UPDATE 시 기존에 DB유저와 파라미터 유저 비교 후 없으면 기본 일반 그룹으로 GRP_ID 수정
-					if(userList.size() > 0) {
-						deleteOprtrIdList = userList.stream().filter(x -> !updateOprtrIdList.contains(x.getOprtrId())).map(x -> x.getOprtrId()).collect(Collectors.toList());
-					}
-				}
+			MOpGrpInfo defaultGeneralGrpInfo = mOpGrpInfoMapper.findOneMOpGrpInfoForDefaultGrp(defaultGrpInfo);
+			if(defaultGeneralGrpInfo == null) {
+				throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"일반 관리자 기본 그룹이 존재하지 않습니다. 생성 후 삭제해 주세요.");
 			}
 			
-			//삭제할 목록이 존재하면 MOpOperator Update
-			if(deleteOprtrIdList.size() > 0) {
-				//기본 그룹, 일반 관리자 그룹 DB 조회
-				MOpGrpInfo defaultGrpInfo = new MOpGrpInfo();
-				defaultGrpInfo.setBscGrpYn("Y");//기본 그룹
-				defaultGrpInfo.setUpperOprtrAuthGrpYn("N");//일반 관리자 그룹
-				
-				MOpGrpInfo defaultGeneralGrpInfo = mOpGrpInfoMapper.findOneMOpGrpInfoForDefaultGrp(defaultGrpInfo);
-				if(defaultGeneralGrpInfo == null) {
-					throw new CommonException(ErrorCode.ENTITY_DATA_NULL,"일반 관리자 기본 그룹이 존재하지 않습니다. 생성 후 삭제해 주세요.");
-				}
-				
-				Map<String,Object>  mOpOperatorUpdateMap = new HashMap<String, Object>();
-				mOpOperatorUpdateMap.put("updateOprtrIdList", deleteOprtrIdList);
-				mOpOperatorUpdateMap.put("grpId", defaultGeneralGrpInfo.getGrpId());
-				mOpOperatorMapper.updateForGrpInfo(mOpOperatorUpdateMap);
-			}
-			
-		} catch (Exception e) {
-			throw new CommonException(ErrorCode.ENTITY_UPDATE_FAIL,"유저 그룹정보가 수정되지 않았습니다.");
+			Map<String,Object>  mOpOperatorUpdateMap = new HashMap<String, Object>();
+			mOpOperatorUpdateMap.put("updateOprtrIdList", deleteOprtrIdList);
+			mOpOperatorUpdateMap.put("grpId", defaultGeneralGrpInfo.getGrpId());
+			mOpOperatorMapper.updateForGrpInfo(mOpOperatorUpdateMap);
 		}
 	}
 }
